refactor : ProductItem 안의 태그들을 컴포넌트로 분리 및 context를 사용해 각 컴포넌트에서 데이터를 받아오도록 수정
